pyproject.toml
[tool.poetry]
name = "dos-locos"
version = "0.1.0"
description = ""
authors = ["Rajat Jacob <rajatjacob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.37.2"
torch = "^2.2.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
